# src/Samples/Liftr.Sample.Web.Tests/expecting-swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
    "/api/MarketplaceResource/{resourceId}/{marketplaceSubId}": {
      "put": {
        "tags": [
          "MarketplaceResource"
        ],
        "operationId": "Put",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "resourceId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "marketplaceSubId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Default error response.",
            "schema": {
              "$ref": "#/definitions/ResourceProviderDefaultErrorResponse"
            }
          }
        },
        "x-ms-examples": {
          "Put": {
            "$ref": "./examples/Put.json"
          }
        }
      }
    },
    "/api/MarketplaceResource/{marketplaceSubId}": {
      "get": {
        "tags": [
          "MarketplaceResource"
        ],
        "operationId": "Get",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "marketplaceSubId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Default error response.",
            "schema": {
              "$ref": "#/definitions/ResourceProviderDefaultErrorResponse"
            }
          }
        },
        "x-ms-examples": {
          "Get": {
            "$ref": "./examples/Get.json"
          }
        }
      }
    },
    "/api/MarketplaceResource/billing/listsaasresources": {
      "get": {
        "tags": [
          "MarketplaceResource"
        ],
        "operationId": "Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "type": "string"
          },
          {
            "in": "query",
            "name": "page-size",
            "type": "integer",
            "format": "int32",
            "default":  10
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SaasResourcesListResponse"
            }
          },
          "default": {
            "description": "Default error response.",
            "schema": {
              "$ref": "#/definitions/ResourceProviderDefaultErrorResponse"
            }
          }
        },
        "x-ms-examples": {
          "Get": {
            "$ref": "./examples/Get.json"
          }
        }
      }
    },
    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Datadog/monitors": {
      "get": {
        "tags": [
          "MockResource"
        ],
        "summary": "List all monitors under the specified resource group.",
        "operationId": "Monitors_ListByResourceGroup",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "subscriptionId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "resourceGroupName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DatadogMonitorResourceListResponse"
            }
          },
          "default": {
            "description": "Default error response.",
            "schema": {
              "$ref": "#/definitions/ResourceProviderDefaultErrorResponse"
            }
          }
        },
        "x-ms-pageable": {
          "nextLinkName": "nextLink"
        },
        "x-ms-examples": {
          "Monitors_ListByResourceGroup": {
            "$ref": "./examples/Monitors_ListByResourceGroup.json"
          }
        }
      }
    },
    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Datadog/monitors/{monitorName}": {
      "put": {
        "tags": [
          "MockResource"
        ],
        "summary": "Create a monitor resource.",
        "operationId": "Monitors_Create",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "subscriptionId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "resourceGroupName",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "monitorName",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DatadogMonitorResource"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DatadogMonitorResource"
            }
          },
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DatadogMonitorResource"
            }
          },
          "default": {
            "description": "Default error response.",
            "schema": {
              "$ref": "#/definitions/ResourceProviderDefaultErrorResponse"
            }
          }
        },
        "x-ms-long-running-operation": true,
        "x-ms-examples": {
          "Monitors_Create": {
            "$ref": "./examples/Monitors_Create.json"
          }
        }
      }
    }
  },
  "definitions": {
    "ErrorResponseBody": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "target": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ErrorResponseBody"
          }
        }
      }
    },
    "ResourceProviderDefaultErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/ErrorResponseBody"
        }
      }
    },
    "SaasBeneficiary": {
      "type": "object",
      "properties": {
        "emailId": {
          "type": "string"
        },
        "objectId": {
          "type": "string"
        },
        "tenantId": {
          "type": "string"
        },
        "puid": {
          "type": "string"
        }
      }
    },
    "SubscriptionTerm": {
      "type": "object",
      "properties": {
        "termUnit": {
          "type": "string"
        },
        "startDate": {
          "format": "date-time",
          "type": "string"
        },
        "endDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "SaasSubscriptionStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer",
      "x-ms-enum": {
        "modelAsString": true,
        "name": "SaasSubscriptionStatus"
      }
    },
    "SaasAdditionalMetadata": {
      "type": "object",
      "properties": {
        "azureSubscriptionId": {
          "type": "string"
        }
      }
    },
    "MarketplaceSubscriptionDetailsEntity": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "planId": {
          "type": "string"
        },
        "offerId": {
          "type": "string"
        },
        "publisherId": {
          "type": "string"
        },
        "beneficiary": {
          "$ref": "#/definitions/SaasBeneficiary"
        },
        "term": {
          "$ref": "#/definitions/SubscriptionTerm"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "saasSubscriptionStatus": {
          "$ref": "#/definitions/SaasSubscriptionStatus"
        },
        "additionalMetadata": {
          "$ref": "#/definitions/SaasAdditionalMetadata"
        }
      }
    },
    "SaasResourcesListResponse": {
      "type": "object",
      "properties": {
        "subscriptions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MarketplaceSubscriptionDetailsEntity"
          }
        },
        "nextLink": {
          "type": "string"
        }
      }
    },
    "DatadogMonitorResource": {
      "type": "object",
      "properties": {
        "id": {
          "description": "ARM id of the monitor resource.",
          "type": "string",
          "readOnly": true
        },
        "name": {
          "description": "Name of the monitor resource.",
          "type": "string",
          "readOnly": true
        },
        "type": {
          "description": "The type of the monitor resource.",
          "type": "string",
          "readOnly": true
        },
        "tags": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "location": {
          "type": "string"
        }
      }
    },
    "DatadogMonitorResourceListResponse": {
      "type": "object",
      "properties": {
        "value": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DatadogMonitorResource"
          }
        },
        "nextLink": {
          "type": "string"
        }
      }
    }
  }
}